Extract screenshot capture and strip fonts
